Move grant config to root
